#145948067: legacy id to 19-digit code

+++ src/main/java/gov/ca/cwds/data/persistence/cms/Base62.java
@@ -26,7 +26,7 @@ public class Base62 {
     return i / BASE;
   }
 
-  public static int toBase10(String str) {
+  public static long toBase10(String str) {
     return toBase10(new StringBuilder(str).reverse().toString().toCharArray());
   }
